Add connection failure handler
RTCPeerConnection.onicecandidate(event)
receives a null candidate in the event
on connection state change,
so when this happens,
check connection state
and call the peer handler
to close the connection
and alert the user
when state changes to "failed".

This alerts only the peer that fails ICE,
so the other peer has no idea why the connection closes.