Remove facingMode restriction.
I added this restriction because
the new Firefox Android was not allowing
camera selection and was automatically selecting
the rear-facing camera.
Considering the current state of Mozilla,
I guessed that this was the new normal,
but things appear worse than I had assumed
so I'm going to ignore Firefox Android for now
and assume that this problem
will eventually get worked out
with the rest of them
or go away when Firefox does.