Improve connect/disconnect handling.
Bubble connect/disconnect events
up from the websocket,
through the client
to connect/disconnect callbacks
set on the client by controllers.

Show an "offline" label in the nav menu
when the peer object is created.
Populate the nav menu
with the peers button on websocket connect
and revert to the offline label
on websocket disconnect.

Add a reset method to the peer dialog
to clear lists on websocket disconnect.

Add a halt method to the websocket
to stop the websocket from re-connecting
on punt (and otherwise).

Call the websocket halt method
in the client's disconnect method
when the websocket isn't connected.
Previously,
if the websocket was in a connect-retry state
the client kept trying to connect
even after disconnect was called.

Change the name of the websocket isDisconnecting variable
to isHalted to make its purpose clearer.
Reset the variable to false
in the websocket connect method
instead of in the disconnect method
to avoid setting isHalted to true permanently
when disconnect is called
when the websocket isn't connected.

In the websocket disconnect method,
set the socket.onclose event
in the socket existence condition
instead of in a separate isConnected condition.
Since the isConnected condition
implies socket existence,
this should work the same as before
but it's easier for me to understand
since all socket handlers
are set together in the same clause
instead of separately.