Halt connection on ping timeout.
Vivaldi on Android 6
slows the keepalive `setInterval` timer
when it hibernates so that the client
doesn't send keepalive pings
before the browser disconnects the client,
as usual, after a minute of inactivity.
However, the client reconnects the WebSocket
when the browser times out the WebSocket.

Chrome on Android 10
seems to do the same thing
when the tab is in the background,
but Firefox, Vivaldi and Chromium on Debian 10
run `setInterval` normally in background tabs.

This change halts the WebSocket connection
when the client hasn't sent the keepalive ping
before the browser disconnects the WebSocket.

It's probably not a great idea to do this
since I'm not sure what actual problem it solves,
but here it is.