Add fail message to connection dialog protocol.
One peer sends fail instead of close
in situations where the peers attempt a connection,
but one peer fails to marshal
a resource necessary for the connection.

The accepting peer sends fail
when it accepts the offer,
but fails to initialize local media,
and both offering and accepting peers send fail
after successfully initializing local media
but upon failing to find a suitable ICE candidate.

The failing peer, as previously,
shows an alert dialog that indicates
the exact nature of the failure.

The peer that receives the fail message
also shows an alert dialog,
but the exact nature of the failure
is unknown to it.

Change ICE failure logs to error level.