Show peers in the main content window.
Split PeersDialog into PeersPanel and OffersDialog.

Improve modal/alert handling.

Review and sanitize
offer/cancel/accept/close handling,
particularly of UI components
when closing connections.

Remove the "Peers" nav menu item
and show an "Online" label instead.

Add a status message to the PeersPanel
that's visible only when it contains no peers.
Show "Offline" when the client is disconnected
and "Waiting for peers" when connected.

Hide the PeersPanel when a connection opens
and unhide it when it closes.

Show the video element when a connection opens
and hide it when the it closes.

Set the video element's background color to black
and it's object-fit style to "contain" at all times.

Add an ignore button to OffersDialog
that removes all peers from the dialog,
but doesn't send close to the offering peers.

OffersDialog lists all offering peers
and is dismissed by accepting an offer,
by ignoring all current offers,
or when offering peers close connections
and no more offers are available.

The dialog defines a hasContent method
that returns true when it contains active offers.
The view's AlertModal uses this method
to re-show the OffersDialog
when the alert is dismissed.

Since the view's showDialog now expects
an object with a getContent method as an argument,
replace the _offerDialog method
with the OfferingDialog class.

Set the OfferingDialog's message
to "Starting local media"
and show it as soon as an offer is made.
Change the message to "Offer sent..."
after local media initializes successfully
and the offer is sent.

Change the view's showModal method to accept an object
and hide the setModalContent method.
The modal object must define a getContent method
and can optionally define a hasContent method.

The getContent method returns a list of
the dialog's header/section/footer elements.

The view keeps a reference
to the active modal object,
and a non-null value indicates that
a modal is active.

The view's showAlert method
replaces modal content
with the AlertDialog's elements,
but stores the current modal object
so it can show its elements again
when the alert is dismissed.

When the replaced modal defines a hasContent method,
the AlertDialog calls the method when it's dismissed
and re-shows the old dialog's elements
when its hasContent method returns something truthy.

This is useful when the OffersDialog
contains more than one offering peer,
but accepting an offer generates an alert.
If the OffersDialog has more offers in its list
when the AlertDialog is dismissed,
the view re-shows the OffersDialog.

By default, modals disable modal control,
so modals must provide a close button
or some some other method of closing the modal.

To make the modal dismissable,
modal objects can set a truthy value
on an optional enableCloseControl variable,
and they can add a close button
to the header returned by getModalHeader
by passing true to the method's second argument.