Improve echo timer.
Change echo timer from interval to timer
and set a new timer on every ping.
Interval and timer seem to work the same,
slowing down and stopping
when the browser is put in the background,
or when the Android device goes to sleep,
but a timer works better
with the focus listener.

Add a focus listener
that clears the existing timer
and sends an echo on every focus event.
This stops the client from disconnecting
when it foregrounds or wakes up,
but the timer has slowed enough
that the echo doesn't happen
before the browser times out the WebSocket.