Add connection label.
When a connection is active,
show a label in nav status describing the connected peer
instead of showing the name button and local peer.

Add peerName to Peer._onAcceptOffer
and change OffersDialog.onAccept to send the offer's peerName
as well as clientId.

Add peerName property to accept message event data,
and add (validated) peerName to Peer._handleAccept's signature.

Add peerName to Peer._openConnection arguments
and update the method to initialize the connection label
and add it to nav status.

Call Peer._closeConnection
in Peer._onAcceptOffer and Peer._openConnection.
This improves DRY and localizes the code
that changes both nav status and nav menu content
when connections open and close.

Remove unused Connection constructor variable
and fix isConnectedTo to require clientId argument
and to return false when the connection has been initialized,
but not for the given clientId.