Review and update how connections open.
Moved RTCPeerConnection handlers
from Connection.open arguments
to empty-function class variables.

Peer sets callbacks on it's Connection object
before opening connections.

This has the nice effect
of bringing ICE failure handling
together into a single function.

Fix connection failure handling.

A connection is considered open
when a peer has either accepted a connection offer
or received an accept message in reply to an offer.

Open connections can fail
either when the accepting peer
can't initialize user media
or when either peer fails ICE.

In either case,
when a client fails to open a connection
it closes the connection,
shows an alert containing details of the error
and sends an error message to the other peer.
When a client receives an error message
from its connected peer,
it closes the connection
and shows an alert that explains
that the other peer failed to connect,
but that contains no deatils of the error.