Add offline-trying peer state.
Don't call Client.onDisconnect
in Client.disconnect
so that onDisconnect only gets called
when the WebSocket disconnects.
It's too confusing
to manage in Peer
when it gets called
from two different places.

Add a method to Peer
that sets view state
based on client connection state
and modify Peer's
client control methods and callbacks
to accommodate.